feat: pr 선택 시 글쓰기 페이지로 이동
